Add/Remove some endpoints.
GetCharmURLOrigin is removed. An implementation detail necessary for application
deploy and refresh with the juju 3.x clients. Should not be need in the new api
definition.

Added Resource Attach to mimic the `juju attach-resource` command.

Removed Payload List, Pinger Ping, and Pinger Stop. They are used by juju agents
and not the actual juju client.